# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API untuk platform ujian online IPPAT (Fiber + GORM + PostgreSQL). Semua response dibungkus format {success, message, data, errors}.\nEndpoint yang butuh login (guru/admin/peserta) ditandai kunci gembok di UI ini — klik \"Authorize\" dan isi `Bearer \u003ctoken\u003e` yang didapat dari /api/auth/login atau /api/auth/register.",
        "title": "IPPAT Backend API",
        "contact": {
            "name": "IPPAT Backend Team"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/kelas": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kelas"
                ],
                "summary": "List kelas",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Nomor halaman",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Jumlah data per halaman",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter berdasarkan ID Jurusan",
                        "name": "id_jurusan",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter berdasarkan tingkat",
                        "name": "tingkat",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_kelas_dto.KelasListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kelas"
                ],
                "summary": "Buat kelas baru",
                "parameters": [
                    {
                        "description": "Data kelas",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_internal_modules_kelas_dto.CreateKelasRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_kelas_dto.KelasResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
        "/kelas/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kelas"
                ],
                "summary": "Detail kelas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Kelas",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_kelas_dto.KelasResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kelas"
                ],
                "summary": "Update kelas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Kelas",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data kelas",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_internal_modules_kelas_dto.UpdateKelasRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_modules_kelas_dto.KelasResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kelas"
                ],
                "summary": "Hapus (soft delete) kelas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Kelas",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
        "/kelas/{id}/restore": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kelas"
                ],
                "summary": "Pulihkan kelas yang sudah dihapus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Kelas",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_helpers.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "backend_internal_helpers.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "backend_internal_modules_kelas_dto.CreateKelasRequest": {
            "type": "object",
            "required": [
                "id_jurusan",
                "nama_kelas",
                "tingkat"
            ],
            "properties": {
                "id_jurusan": {
                    "type": "string"
                },
                "nama_kelas": {
                    "type": "string"
                },
                "tingkat": {
                    "type": "string"
                }
            }
        },
        "backend_internal_modules_kelas_dto.KelasListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend_internal_modules_kelas_dto.KelasResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "backend_internal_modules_kelas_dto.KelasResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "id_jurusan": {
                    "type": "string"
                },
                "nama_jurusan": {
                    "type": "string"
                },
                "nama_kelas": {
                    "type": "string"
                },
                "tingkat": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "backend_internal_modules_kelas_dto.UpdateKelasRequest": {
            "type": "object",
            "required": [
                "id_jurusan",
                "nama_kelas",
                "tingkat"
            ],
            "properties": {
                "id_jurusan": {
                    "type": "string"
                },
                "nama_kelas": {
                    "type": "string"
                },
                "tingkat": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Ketik \"Bearer\" diikuti spasi dan JWT token. Contoh: \"Bearer eyJhbGciOi...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}